feature/37 refactor name
