+++ app/src/widgets/Message/hooks/useMessageVariant.ts
