--- packages/ui-kit/MessageItem/MessageItem.module.scss
@@ -1,7 +1,0 @@
-.content {
-  display: flex;
-}
-
-.primary {
-  flex-direction: row-reverse;
-}
