--- packages/ui-kit/MessageItem/MessageItem.tsx
@@ -1,74 +1,0 @@
-import { FC, useMemo } from 'react';
-import cn from 'classnames';
-
-import { ActionItem, ActionsDisplay } from '../Actions';
-import { Actions } from '../index';
-import { PopoverPosition, PopoverSide } from '../Popover';
-import { TextViewer, TextViewerVariant } from '../TextViewer';
-
-import { MessageItemVariant } from './types';
-
-import css from './MessageItem.module.scss';
-
-export interface MessageItemProps {
-    textTextViewer: string;
-    actions: ActionItem[];
-    countShowActions?: number;
-    positionActions?: PopoverPosition;
-    sideActions?: PopoverSide;
-    titleTextViewer?: string;
-    subTextTextViewer?: string;
-    variant?: MessageItemVariant;
-}
-
-const MessageItem: FC<MessageItemProps> = ({
-    actions,
-    countShowActions,
-    positionActions,
-    sideActions,
-    titleTextViewer,
-    textTextViewer,
-    subTextTextViewer,
-    variant = MessageItemVariant.PRIMARY,
-}) => {
-    const { variantTextViewer, displayActions } = useMemo(() => {
-        let vVariant: TextViewerVariant;
-        let aActions: ActionsDisplay;
-
-        switch (variant) {
-            case MessageItemVariant.PRIMARY:
-                vVariant = TextViewerVariant.PRIMARY;
-                aActions = ActionsDisplay.RIGHT;
-                break;
-            case MessageItemVariant.SECONDARY:
-                vVariant = TextViewerVariant.SECONDARY;
-                aActions = ActionsDisplay.LEFT;
-                break;
-            default:
-                vVariant = TextViewerVariant.PRIMARY;
-                aActions = ActionsDisplay.RIGHT;
-        }
-
-        return { variantTextViewer: vVariant, displayActions: aActions };
-    }, [variant]);
-
-    return (
-        <div className={cn(css.content, css[variant])}>
-            <TextViewer
-                text={textTextViewer}
-                title={titleTextViewer}
-                variant={variantTextViewer}
-                subText={subTextTextViewer}
-            />
-            <Actions
-                actions={actions}
-                countShow={countShowActions}
-                display={displayActions}
-                position={positionActions}
-                side={sideActions}
-            />
-        </div>
-    );
-};
-
-export default MessageItem;
