--- packages/ui-kit/MessageItem/index.ts
@@ -1,2 +1,0 @@
-export { default as MessageItem } from './MessageItem';
-export * from './types';
