--- packages/ui-kit/MessageItem/types.ts
@@ -1,4 +1,0 @@
-export enum MessageItemVariant {
-    PRIMARY = 'primary',
-    SECONDARY = 'secondary',
-}
